(refactor): move test files into test/
- app root has a lot of files in it, this should help declutter a bit
